Struct key_module::driver::Driver_trait::Driver_Interface

#[repr(C)]
pub struct Driver_Interface(_);

A marker type describing the traits that are required when constructing Driver_TO,and are then implemented by it, by implementing the InterfaceType trait.

Implementations

impl Driver_Interface

pub const NEW: Self = _

Constructs this type