Struct key_module::driver::Driver_TO

#[repr(transparent)]
pub struct Driver_TO<'lt, _ErasedPtr>where
    _ErasedPtr: __GetPointerKind,{
    pub obj: Driver_Backend<'lt, _ErasedPtr>,
    /* private fields */
}

The trait object for Driver.

There are extra methods on the obj field.

Fields
obj: Driver_Backend<'lt, _ErasedPtr>

Implementations

impl<'lt, _ErasedPtr> Driver_TO<'lt, _ErasedPtr>where
_ErasedPtr: AsPtr<PtrTarget = ()>,

pub fn from_ptr<_OrigPtr, Downcasting>(
ptr: _OrigPtr,
can_it_downcast: Downcasting
) -> Selfwhere
_OrigPtr: CanTransmuteElement<(), TransmutedPtr = _ErasedPtr>,
_OrigPtr::PtrTarget: Driver + Sized + 'lt,
Driver_Interface: GetRObjectVTable<Downcasting, _OrigPtr::PtrTarget, _ErasedPtr, _OrigPtr>,

Constructs this trait object from a pointer to a type that implements Driver.

This method is automatically generated, for more documentation you can look at [abi_stable::docs::sabi_trait_inherent#from_ptr-method]

can_it_downcast describes whether the trait object can be converted back into the original type or not.
Its possible values are TD_CanDowncast and TD_Opaque.

pub fn from_sabi(obj: Driver_Backend<'lt, _ErasedPtr>) -> Self

Constructs this trait object from its underlying implementation.

This method is automatically generated, for more documentation you can look at [abi_stable::docs::sabi_trait_inherent#from_sabi-method]

pub fn sabi_reborrow<'_sub>(&'_sub self) -> Driver_TO<'lt, RRef<'_sub, ()>>

Reborrows this trait object to a reference-based trait object.

This method is automatically generated, for more documentation you can look at [abi_stable::docs::sabi_trait_inherent#sabi_reborrow-method]

pub fn sabi_reborrow_mut<'_sub>(
&'_sub mut self
) -> Driver_TO<'lt, RMut<'_sub, ()>>where
_ErasedPtr: AsMutPtr<PtrTarget = ()>,

Reborrows this trait object to a mutable-reference-based trait object.

This method is automatically generated, for more documentation you can look at [abi_stable::docs::sabi_trait_inherent#sabi_reborrow_mut-method]

impl<'lt> Driver_TO<'lt, RBox<()>>

pub fn from_value<_Self, Downcasting>(
ptr: _Self,
can_it_downcast: Downcasting
) -> Selfwhere
_Self: Driver + 'lt,
Driver_Interface: GetRObjectVTable<Downcasting, _Self, RBox<()>, RBox<_Self>>,

Constructs this trait from a type that implements Driver.

This method is automatically generated, for more documentation you can look at [abi_stable::docs::sabi_trait_inherent#from_value-method]

can_it_downcast describes whether the trait object can be converted back into the original type or not.
Its possible values are TD_CanDowncast and TD_Opaque.

impl<'lt, '_sub> Driver_TO<'lt, RRef<'_sub, ()>>

pub const fn from_const<_Self, Downcasting>(
ptr: &'_sub _Self,
can_it_downcast: Downcasting
) -> Selfwhere
_Self: Driver + 'lt,
_Self: 'lt,
Driver_Interface: GetRObjectVTable<Downcasting, _Self, RRef<'_sub, ()>, &'_sub _Self>,

Constructs this trait from a constant of a type that implements Driver.

This method is automatically generated, for more documentation you can look at [abi_stable::docs::sabi_trait_inherent#from_const-method]

You can construct the vtable_for parameter with [Driver_MV::VTABLE].

can_it_downcast describes whether the trait object can be converted back into the original type or not.
Its possible values are TD_CanDowncast and TD_Opaque.

impl<'lt, _ErasedPtr> Driver_TO<'lt, _ErasedPtr>where
_ErasedPtr: AsPtr<PtrTarget = ()>,

pub fn load_data<'borr>(&mut self, param_0: RString) -> RResult<u64, RString>where
_ErasedPtr: AsMutPtr<PtrTarget = ()>,

This is the inherent equivalent of the trait method of the same name

pub fn poll(&mut self, param_0: u64) -> RResult<RVec<u16>, RString>where
_ErasedPtr: AsMutPtr<PtrTarget = ()>,

This is the inherent equivalent of the trait method of the same name

pub fn set(
&mut self,
param_0: u64,
param_1: usize,
param_2: u16
) -> RResult<(), RString>where
_ErasedPtr: AsMutPtr<PtrTarget = ()>,

This is the inherent equivalent of the trait method of the same name