Struct key_module::hid::HidModuleRef

#[repr(transparent)]
pub struct HidModuleRef(pub PrefixRef<HidModule_Prefix>);

This is the pointer to the prefix of HidModule.

This is automatically generated documentation,by the StableAbi derive macro.

Creating a compiletime-constant

You can look at the docs in [::abi_stable::docs::prefix_types] to see how you can construct and use this and similar types.
More specifically in the “constructing a module” example or the “Constructing a vtable” example

Tuple Fields
0: PrefixRef<HidModule_Prefix>